+++ src/Raven.CodeAnalysis/Symbols/ISymbol.cs
@@ -494,10 +494,7 @@ public interface ITypeSymbol : INamespaceOrTypeSymbol
             if (this is IUnionCaseTypeSymbol)
                 return true;
 
-            return GetAttributes().Any(static attribute =>
-                attribute.AttributeClass?.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat) is
-                    "System.Runtime.CompilerServices.UnionCaseAttribute" or
-                    "System.Runtime.CompilerServices.DiscriminatedUnionCaseAttribute");
+            return false;
         }
     }
 
@@ -510,19 +507,6 @@ public interface ITypeSymbol : INamespaceOrTypeSymbol
 
             if (this is IUnionCaseTypeSymbol caseSymbol)
                 return caseSymbol.Union;
-
-            foreach (var attribute in GetAttributes())
-            {
-                if (attribute.AttributeClass?.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat) is
-                        "System.Runtime.CompilerServices.UnionCaseAttribute" or
-                        "System.Runtime.CompilerServices.DiscriminatedUnionCaseAttribute" &&
-                    attribute.ConstructorArguments.Length == 1)
-                {
-                    var argument = attribute.ConstructorArguments[0];
-                    if (argument.Kind == TypedConstantKind.Type && argument.Value is INamedTypeSymbol unionType)
-                        return unionType;
-                }
-            }
 
             return null;
         }